python/requirements.txt
pandas==1.5.3
numpy==1.24.3
statsmodels==0.13.5
mysql-connector-python==8.0.32
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2 